requirements.txt
Django==3.0.8
django-widget-tweaks==1.4.8
djangorestframework==3.11.0